% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FileNrow}
\alias{FileNrow}
\title{Row number count}
\usage{
FileNrow(filename)
}
\arguments{
\item{filename, }{name of a file.}
}
\description{
To count number of row of a file
}
\examples{
# reading data
bim_path <- system.file("extdata", "ref_geno.bim", package = "SumTool")
n <- FileNrow(bim_path)
}
