% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumTool.r
\name{LDcal}
\alias{LDcal}
\title{LD calculation}
\usage{
LDcal(geno = NULL, index = NULL, threads = 1, lambda = 0,
  chisq = 0, correlation = TRUE, out = NULL, verbose = TRUE)
}
\arguments{
\item{geno}{big.matrix (n * m), genotype}

\item{index}{vector, only calculate LD for a subset with all SNPs (m * m1)}

\item{threads}{number, the number of used threads for parallel process}

\item{lambda}{number, add a value on the diagonal of LD matrix}

\item{chisq}{number, generate a sparse LD matrix, if ld^2 * n < chisq, it will be set to 0}

\item{correlation}{logical, if TRUE, the LD matrix will be constructed by correlation of all pairs}

\item{out}{character, the path and prefix of output file, if not NULL, the LD matrix will stored in big.matrix, which will save much memory, but cost a bit more time}

\item{verbose}{logical, whether to print the log information}
}
\description{
To calculate LD for all pairs of SNPs on given genome
}
\examples{
gwas_bfile_path <- system.file("extdata", "gwas_geno", package = "SumTool")
data <- read_plink(bfile=gwas_bfile_path, threads=1)
geno <- data$geno
ld <- LDcal(geno=geno, threads=1)
}
