% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumTool.r
\name{LDclump}
\alias{LDclump}
\title{LD clumping}
\usage{
LDclump(
  geno = NULL,
  map = NULL,
  p = NULL,
  w = 1e+06,
  r2.cutoff = 0.25,
  p.cutoff = 1,
  verbose = TRUE,
  threads = 1
)
}
\arguments{
\item{geno}{bigmemory (n * m), genotype coded as 0, 1, 2}

\item{map}{data.frame, the genomic information of SNPs. The columns should be in the order of c("SNP", "Chr", "Pos", "A1", "A2")}

\item{p}{data.frame, at least 2 columns, the first column should be SNP names, the last column should be the pvalues}

\item{w}{int, size of windows in bp. Default is 1e6}

\item{r2.cutoff}{double, the threshold of r2, smaller cutoff results in less remaining SNPs, default 0.25}

\item{p.cutoff}{double, the threshold of p-values, smaller threshold results in less remaining SNPs, default 1}

\item{verbose}{logical, whether to print the log information}

\item{threads}{int, the number of used threads for parallel process}
}
\description{
To remove high correlated SNPs on base of r2 and p-values
}
\examples{
ref_bfile_path <- system.file("extdata", "ref_geno", package = "SumTool")
p_path <- system.file("extdata", "P.txt", package = "SumTool")
# load data
data <- read_plink(bfile=ref_bfile_path, threads=1, backingpath=tempdir(), verbose=FALSE)
geno <- data$geno
map <- data$map
pdata <- read.table(p_path, header = TRUE)
snp <- LDclump(geno=geno, map=map, p=pdata, p.cutoff=1, r2.cutoff=0.25, w=100000, verbose=FALSE)
}
