% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumTool.r
\name{LDreg}
\alias{LDreg}
\title{LD regression}
\usage{
LDreg(
  sumstat = NULL,
  ldscore = NULL,
  wld = NULL,
  M = NULL,
  maxz2 = 80,
  maf = 0.05,
  nblock = 200,
  verbose = TRUE
)
}
\arguments{
\item{sumstat}{data.frame or list, results of single trait or multiple traits of GWAS summary statistics. For each summary statistic of a trait, the columns should be in the order of c("SNP", "Chr", "Pos", "A1", "A2", "BETA", "SE", "N"). If it's a data.frame, the heritability will be estimated, if it's a list containing multiple GWAS summary statistics, both heritability of traits and the genetic correlation of pair of traits will be estimated.}

\item{ldscore}{data.frame, the ldscore for each SNP, the columns should be prepared in order of c("SNP", "Chr", "Pos", "A1", "A2", "MAF", "ldscore")}

\item{wld}{data.frame, at least 2 colmuns, the first column is th SNP names, the last column is the weight for each SNP. Default equals to the 'ldscore' option}

\item{M, }{integer, the number of total SNPs that are used to calculate LDscore}

\item{maxz2}{double, max chi^2. Default is 80}

\item{maf}{double, Minor allele frequency lower bound. Default is MAF > 0.05}

\item{nblock}{int, number of blocks for jackknife estimation}

\item{verbose}{logical, whether to print the log information}
}
\description{
To estimate heritability or genetic correlation
}
\examples{
# single trait 
sumstat1_path <- system.file("extdata", "sumstat1", package = "SumTool")
ldscore_path <- system.file("extdata", "ldscore", package = "SumTool")
sumstat1 <- read.table(sumstat1_path, header=TRUE)
ldscore <- read.table(ldscore_path, header=TRUE)
res1 <- LDreg(sumstat = sumstat1, ldscore = ldscore, verbose=FALSE)

#multiple traits
sumstat2_path <- system.file("extdata", "sumstat2", package = "SumTool")
sumstat2 <- read.table(sumstat2_path, header=TRUE)
res2 <- LDreg(sumstat = list(sumstat1, sumstat2), ldscore = ldscore, verbose=FALSE)
}
