% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumTool.r
\name{LDsore}
\alias{LDsore}
\title{LD score calculation}
\usage{
LDsore(geno = NULL, map = NULL, w = 1e+06, b = 5e+05, r2 = TRUE,
  adjust = TRUE, verbose = TRUE, threads = 1)
}
\arguments{
\item{geno}{bigmemory (n * m), genotype coded as 0, 1, 2}

\item{map}{data.frame, the genomic information of SNPs. The columns should be in the order of c("SNP", "Chr", "Pos", "A1", "A2")}

\item{w}{int, size of windows in bp. Default is 1e6}

\item{b}{number, set the buffer size  in bite of each window, default 500000}

\item{r2}{logical, calculate r2 or r}

\item{adjust}{logical, whether to adjust the ldscore}

\item{verbose}{logical, whether to print the log information}

\item{threads}{number, the number of used threads for parallel process}
}
\description{
To estimate LD score for each SNPs
}
\examples{
ref_bfile_path <- system.file("extdata", "ref_geno", package = "SumTool")

# load data
data <- read_plink(bfile=ref_bfile_path, threads=1)
geno <- data$geno
map <- data$map
ldscore <- LDsore(geno = geno, map = map, w = 100000, b=50000, threads = 1)
}
