% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumTool.r
\name{LMreg}
\alias{LMreg}
\title{Generalized Linear Model}
\usage{
LMreg(y, geno = NULL, map = NULL, X = NULL, threads = 1,
  verbose = TRUE)
}
\arguments{
\item{y}{vector, the phenotype vector with length n}

\item{geno}{big.matrix (n * m), genotype}

\item{map}{data.frame, the genomic information of SNPs. The columns should be in the order of c("SNP", "Chr", "Pos", "A1", "A2")}

\item{X}{matrix, covariates}

\item{threads}{number, the number of used threads for parallel process}

\item{verbose}{logical, whether to print the log information}
}
\description{
To do association using Generalized Linear Model
}
\examples{
ref_bfile_path <- system.file("extdata", "ref_geno", package = "SumTool")
data <- read_plink(bfile=ref_bfile_path, threads=1)
geno <- data$geno
map <- data$map
y <- data$pheno[,1]
gwas <- LMreg(y=y, geno=geno, map=map, threads=1, verbose=FALSE)
}
