% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumTool.r
\name{read_plink}
\alias{read_plink}
\title{Data reading}
\usage{
read_plink(bfile = "", maxLine = 1000, impute = TRUE,
  backingpath = NULL, descriptorfile = NULL, backingfile = NULL,
  verbose = TRUE, threads = 1)
}
\arguments{
\item{bfile}{character, prefix of Plink binary format data.}

\item{maxLine}{number, set the number of lines to handle at a time, bigger lines require more memory.}

\item{impute}{logical, whether to impute missing values in genotype.}

\item{backingpath}{the path to the directory containing the file backing cache.}

\item{descriptorfile}{the name of the file to hold the backingfile description, for subsequent use with ‘attach.big.matrix’; if ‘NULL’, the ‘backingfile’ is used as the root part of the descriptor file name.  The descriptor file is placed in the same directory as the backing files.}

\item{backingfile}{the root name for the file(s) for the cache.}

\item{verbose}{logical, whether to print the log information}

\item{threads}{number, the number of used threads for parallel process}
}
\description{
To read Plink binary format data into bigmemory (0, 1, 2) format
}
\examples{
ref_bfile_path <- system.file("extdata", "ref_geno", package = "SumTool")
data <- read_plink(bfile=ref_bfile_path, threads=1)
geno <- data$geno
map <- data$map
}
