% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumTool.r
\name{read_vcf}
\alias{read_vcf}
\title{Data reading}
\usage{
read_vcf(
  vfile = "",
  maxLine = 1000,
  impute = TRUE,
  additive = TRUE,
  out = NULL,
  verbose = TRUE,
  threads = 1
)
}
\arguments{
\item{vfile}{character, full name of VCF file.}

\item{maxLine}{number, set the number of lines to handle at a time, bigger lines require more memory.}

\item{impute}{logical, whether to impute missing values in genotype.}

\item{additive}{logical, whether to code the genotype into 0, 1, 2; otherwise, the genotype will be coded into 11, 12, 21, 22, which can be used for summary data imputation.}

\item{out}{character, path and prefix of output file}

\item{verbose}{logical, whether to print the log information}

\item{threads}{number, the number of used threads for parallel process}
}
\description{
To read VCF format data into bigmemory {(0, 1, 2) or (11, 12, 21, 22)} format
}
\examples{
ref_vfile_path <- system.file("extdata", "ref_geno.vcf", package = "SumTool")
data <- read_vcf(vfile=ref_vfile_path, out=tempfile(), threads=1, verbose=FALSE)
geno <- data$geno
map <- data$map
}
