% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumTool.r
\name{SBLUP}
\alias{SBLUP}
\title{Summary data based SNP BLUP}
\usage{
SBLUP(sumstat = NULL, geno = NULL, map = NULL, lambda = NULL,
  w = 1e+06, threads = 1, verbose = TRUE)
}
\arguments{
\item{sumstat}{data.frame, GWAS summary statistics of a trait. The columns should be in the order of c("SNP", "Chr", "Pos", "A1", "A2", "BETA", "SE, "N")}

\item{geno}{bigmemory (n * m), genotype coded as 0, 1, 2}

\item{map}{data.frame, the genomic information of SNPs. The columns should be in the order of c("SNP", "Chr", "Pos", "A1", "A2")}

\item{lambda}{double, the ridge regression coefficient, lambda=m*(1/h2-1), m is the number of SNPs, h2 is the heritability of trait}

\item{w}{int, size of windows in bp. Default is 1e6}

\item{threads}{number, the number of used threads for parallel process}

\item{verbose}{logical, whether to print the log information}
}
\description{
To estimate joint SNP effect using GWAS summary statistics
}
\examples{
sumstat_path <- system.file("extdata", "typed.marginal", package = "SumTool")
ref_bfile_path <- system.file("extdata", "ref_geno", package = "SumTool")

# load data
sumstat <- read.table(sumstat_path, header=TRUE)
data <- read_plink(bfile=ref_bfile_path, threads=1)
geno <- data$geno
map <- data$map
h2 <- 0.5
lambda = nrow(sumstat)*(1/h2-1)
eff <- SBLUP(sumstat=sumstat, geno=geno, map=map, lambda=lambda, threads=1)
}
