% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumTool.r
\name{SImputeZ}
\alias{SImputeZ}
\title{Summary data based imputation on Z-score}
\usage{
SImputeZ(ref.geno = NULL, ref.map = NULL, typed.geno = NULL,
  typed = NULL, w = 1e+06, b = 250000, lambda = NULL,
  maf = 1e-06, correlation = TRUE, verbose = TRUE, threads = 1)
}
\arguments{
\item{ref.geno}{big.matrix (n1 * m1), reference genotype panel}

\item{ref.map}{matrix (m1 * 5): SNPs, Chr, position, A1, A2}

\item{typed.geno}{big.matrix (n2 * m2), individual level genotype for typed SNPs (this file is optional)}

\item{typed}{matrix (m2 * 6): SNPs, Chr, position, A1, A2, Z}

\item{w}{number, set the window size in bite, default 1000000}

\item{b}{number, set the buffer size  in bite of each window, default 250000}

\item{lambda}{number, ridge regression value on LD matrix of typed SNPs: solve(Rtt + diag(lambda))}

\item{maf}{number, SNPs whose minor allele frequency are lower than set value will not be imputed}

\item{correlation}{logical, if TRUE, the LD matrix will be constructed by correlation of all pairs}

\item{verbose}{logical, whether to print the log information}

\item{threads}{number, the number of used threads for parallel process}
}
\description{
To impute Zscore using summary data by SImpute/SImpute-LD
}
\examples{
#----------------SImpute----------------#

# get path of the attached files in package
ref_bfile_path <- system.file("extdata", "ref_geno", package = "SumTool")
typed_z_path <- system.file("extdata", "typed.zscore", package = "SumTool")

# reading data
data <- read_plink(bfile=ref_bfile_path, threads=1)
ref.geno <- data$geno
ref.map <- data$map
typed_z <- read.table(typed_z_path, header=TRUE)

# Impute Zscore
xx <- SImputeZ(ref.geno=ref.geno, ref.map=ref.map, typed=typed_z, threads=1)

#--------------SImpute-LD---------------#

gwas_bfile_path <- system.file("extdata", "gwas_geno", package = "SumTool")
gwas <- read_plink(bfile=gwas_bfile_path, threads=1)
typed.geno <- gwas$geno
typed.map <- gwas$map
#NOTE: the order of SNPs in 'typed.geno' should be consistent with the order in 'typed_z'.
typed.geno <- deepcopy(typed.geno, cols = match(typed_z[, 1], typed.map[, 1]))

# Impute Zscore
xx <- SImputeZ(ref.geno=ref.geno, ref.map=ref.map, typed=typed_z, typed.geno=typed.geno, threads=1)
}
