% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumTool.r
\name{write_plink}
\alias{write_plink}
\title{Data writing}
\usage{
write_plink(geno = NULL, map = NULL, out = "plink", verbose = TRUE,
  threads = 1)
}
\arguments{
\item{geno}{big.matrix, numeric matrix stored in big.matrix format.}

\item{map}{number, set the number of lines to handle at a time, bigger lines require more memory.}

\item{out}{logical, whether to impute missing values in genotype.}

\item{verbose}{logical, whether to print the log information}

\item{threads}{number, the number of used threads for parallel process}
}
\description{
To write bigmemory (0, 1, 2) data into Plink binary format
}
\examples{
# reading data
ref_bfile_path <- system.file("extdata", "ref_geno", package = "SumTool")
data <- read_plink(bfile=ref_bfile_path, threads=1, verbose=FALSE)
geno <- data$geno
map <- data$map

# writing data
# write_plink(geno=geno, map=map, out="./test", threads=1, verbose=FALSE)
}
